Fix for motion_sensor.py error
