+++ custom_components/wyzeapi/wyzeapi/devices/base.py
@@ -2,7 +2,7 @@ class BaseDevice:
     nick_name: str
     product_model: str
     mac: str
-    avaliable: int = 1
+    available: int = 1
 
     def __init__(self, nick_name, product_model, mac):
         self.nick_name = nick_name
